arm64: dts: htc/lge: Remove Easel ION heaps
The ION firmware loading mechanism is wasteful and restrictive. The
reserved Easel ION heap permanently removes 64 MiB of memory from the
AP and limits the maximum combined firmware size (SBL + U-Boot + kernel
+ FDT + ramdisk) to 64 MiB for no reason, resulting in silent corruption
if any slot's payload is too big.

The mnh-sm driver has error handling in place to fall back to "legacy"
firmware loading, which doesn't have any of those issues: it only uses
memory during firmware loading, and allows loading firmware as big as
the endpoint's RAM can handle. It's also the standard way to load
firmware in most other kernel drivers.

It's worth noting that Play Store OTA firmware updates will no longer
work, but that's not an issue because this is an old device that is
unlikely to ever receive an Easel OTA again -- the last OTA was in
2018, back when Easel was still in its infancy. The worst-case scenario
is that an Easel firmware update will be delayed for a month (until the
next security patch is released).

Signed-off-by: Danny Lin <danny@kdrag0n.dev>