misc: mnh: Reduce firmware download DMA request latency
There's no good reason to use a custom loop to wait for the DMA callback
to complete when the kernel has a dedicated completion API that can be
used instead. Waiters will be awakened as soon as the event is fired,
which reduces response time significantly while also reducing CPU usage.

For small DMA requests (e.g. 4-128 KiB), the majority of the firmware
download time is spent sleeping in the callback wait loop instead of
actually performing DMA operations.

Example performance with a 4 KiB buffer:
    Before: avg latency = 4.8 ms, total download time = 55 sec
    After:  avg latency = 198 µs, total download time = 2 sec

4 KiB is rather pessimistic, but a 128 KiB buffer (which can often
happen in real-world usage when memory is fragmented) shows similar
results:
    Before: avg latency = 4.2 ms, total download time = 1.6 sec
    After:  avg latency = 706 µs, total download time = 399 ms

Total firmware download time with a 128 KiB buffer is now down to ~400
ms, which isn't much worse than the ~270 ms time when using a full 4 MB
buffer.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>